requirements.txt
pip-tools
streamlit
pandas
numpy
numba
matplotlib
pyqt
scipy
jupyter
psutil
pillow
opencv
tqdm
django
scikit-learn
sympy
pyserial
scikit-image
scikit-optimize
scikit-learn-intelex
sktime
joblib
imageio
pytorch
torchvision
torchaudio
tensorflow==2.18.*
tensorboardx
crc32c
nltk
spacy
pytesseract
seaborn
selenium
statsmodels
pypdf
librosa
pysoundfile
pylops
mouse
keyboard
datasets
lightning
transformers
fastapi
polars
gdcm
pydicom